Receiver argument with a name should not make fields and methods in scope
Sometimes the docs use the receiver argument's fields without the receiver name, I think this should be allowed but only when no name is given. This allows the programmer not to worry about name collisions (when the receiver is named) which can be confusing. It also seems unnecessary to name the receiver if the fields/methods are not called through that name.