compiler: fix generic function call in `for in` expression
Calling a generic function in a `for in` expression resulted in a C error:

generic_test.tmp.c:4971:2: error: expected expression
 ) array_int tmp2 =  main__get_values ( 1;
 ^

This was due to the `ParserState` struct not saving/restoring `p.cgen.is_tmp`
and also because the code replacing the function name with the generic
counterpart was hard coded to `p.cgen.cur_line` regardless of `is_tmp`.

It already works in `for some_bool(1) {` expressions, since it doesn't
create any temp variables.