Added `init` command. Updated Help file
This commit added the `init` command to v, in addition to the
pre-existing `create` command. It mimics the way Cargo allows the same
commands.

* If a pre-existing `v.mod` file already exists, the initialization is
aborted.
* If a file with the same name as the module already exists, that file is
left as-is.

Instead of creating a separate tool from `create` that has almost the
same functionality, `init` or `create` are options sent to `vcreate.v`.
This required changing `launch_tool` in `simple_tool.v` to accept
additional flags with variadic arguments. In the future, this will allow
other tools to be combined, allowing more code-reuse.