Add `strip_margin` method for strings
This is similar to the Scala method for the same thing. If you have a
multi-line string, you can format it so that you can use the same indentation
levels as the surrounding code.

The default delimiter is `|`, but the user can pick a different at their
desire, with the exception of white-space characters. If a white-space
character is provided, an error message is printed, and the default delimiter
is used instead.