os: rename open_stdin to stdin
Stdin is not opened and cannot be closed. The `open_` prefix implies to the
reader that a resource is being acquired that needs to be cleaned up, but that's
not the case here. In other languages like Go and Rust getting a-hold of stdin
involves accessing a package/module variable without open semantics, and it is
confusing to see the concept of opening stdin here.

Examples of how to access stdin in other languages.

Go:
```go
import "os"

os.Stdin
```

Rust:
```rust
use std::io;

io::stdin()
```