doc: tweak built-in functions section & move `dump` section here
`dump` is mentioned in the interface docs and isn't really an
'advanced' feature. It belongs as a built-in function even though it
has special semantics (just like `it` methods are still methods).

Add `println` subheading.
Make 'Printing custom types' a subheading of `println`.
Change `print` functions to take IStr, not string and note IStr is not
public.
Add `dump(Expr)` to built-in functions list.