vls: use `os.config_dir()` and `os.cache_dir()`
See #17826.

This commit was necessary because ~/.vls clutters the home directory of
users. This commit address the problem by using `os.config_dir()`,
`os.cache_dir()` and `os.vmodules_dir()`. The files are relocated as
follows (showing the defaults for Linux):

    ~/.vls/vls.config.json --> ~/.config/vls/config.json
    ~/.vls/.cache          --> ~/.cache/vls
    ~/.vls/src             --> ~/.vmodules/vls
    ~/.vls/bin/<binfile>   --> ~/.vmodules/vls/bin/<binfile>

The side effects are that the old files in ~/.vls should be moved. This
is done by checking if ~/.vls exists and then moving the files
accordingly before running the program as usual.