vweb: update docstrings for response functions
The docstrings claim that the server will respond with HTTP_OK, but it will actually respond with the most recent set status code. I.e. if you explicitly set the status code before calling one of the response functions, it will use that code and not the default `200 OK` status.