cgen: fix option unwrap for fields of interface type
This commit fixes the codegen for the unwrapping fields of type interface. Without this change, cgen would attempt to type cast the wrapped field to the unwrapped type, resulting in an invalid cast (I_main__AnInterface_as_I_main__AnInterface) causing the C compiler to fail with an implicit declaration error.