js: fix array type checking in sum type match expressions (fix #24237 )
This commit fixes a bug in the JavaScript code generator where array types in sum type match expressions were not being properly checked. The issue was that the generated code was checking if array elements were instances of the array type itself, rather than the element type of the array.