os: add debugger_present implementation for OpenBSD
Check if a child process could trace its parent process, if not a
debugger must be present.

On OpenBSD, impossible to trace a parent process from its child. But
with errors (errno), it's possible to determine if a debugger is present
or not.
Possible errors for ptrace / PT_ATTACH in child process:
  - EBUSY: process already traced
  - EPERM: process is not a child of tracer (sysctl kern.global_ptrace=0)
  - EINVAL: process is an ancestor of the current process and not
    init (sysctl kern.global_ptrace=1)

Signed-off-by: Laurent Cheylus <foxy@free.fr>