thirdparty: Add NDEBUG check before glGetErrorCode if-statement
When built for release, assert(x) becomes NOOP. Commit 37e5ead, had introduced
an if-block & fprintf statement to print the glGetError() code to make
diagnostic easier. However, this change will execute even when assert(x) is
a cast to void(0). Resulting in performance drop - which can be noticable.
(Ex: Using gg with emscripten will result in single-digit framerate)

This commit provides a fix by having an ifdef check for NDEBUG, and if NDEBUG
is defined then _SG_GL_CHECK_ERROR will be defined as nothing, as the result
of SOKOL_ASSERT -> ASSERT, would also be nothing (glibc/assert.h).