os: don't resolve symlinks in find_abs_path_of_executable()
This fixes aliases and wrappers being ignored.

Resolving the symlinks can cause breakage, an example is coreutils:
commands such as "echo" are just a symlink to a single binary.
The name of the symlink tells the program which tool is requested.

Or even worse, in the case of security helpers such as firejail:
a symlink is created in one of the top priority folders in PATH,
so that firejail steps in and knows what to run.
Resolving the symlink would make the program run outside a jail!