gg, examples: add optional `update_fn: fn (dt f32, ctx &gg.Context)` to gg.Context and gg.Config . Pass consistently the current GG context to *all* callback functions, instead of nil.
Previously, when the user_data field: was not set explicitly in the gg.new_context() call,
it was still passed to *some* of the callbacks, which was harder to understand, and hindered simple GG apps.

Now, when the user_data field is not set explicitly, all gg callbacks will receive a pointer to the current gg context, instead of nil.